[fix] getCommittees() was returning wrong values
The list of categories (now renamed ‘roles’) was incomplete (not covering all the possible cases) and the implemented system was too naive.

The refactored code now explicitly tries to identify each committee role and throws an exception if it cannot do so.

Moreover, the structure of the returned array has changed. Now, each key is a committee identifier and each value is an array storing the (possibly multiple) roles of the MP in this committee.