Improve retrieval of seat data
This change makes the method less flexible, since it will not return arrays containing null values anymore when an anchor is not present.

This is done on purpose to prevent members of the Chamber to be returned, since those never have a Senate identifier nor, as a result, an anchor.

It also means that potential senators listed without an anchor linking to their page will not be returned, but it’s not expected to see this kind of things during normal activity periods of the parliament.