[refactor] Simplify retrieval of MP anchors
There is no need to have a really precise CSS selector, nor to use an intermediary Crawler for a table cell. Even if the DOM we deal with is complex, it is supposed to include only a single <a> element.

This commit then removes the `$cell` variable and simplifies the selector. It also updates comments to accommodate these changes.