Add main ‘list’ scraper for K committee agendas
This scraper targets the main page of the committee agenda of the Chamber.

Depending of the content of this page, the scraper instantiates and calls other scrapers in order to get a list of URLs targeting detailed committee agendas.

In other words, this scrapers acts as an ‘intelligent’ intermediary, putting itself in front of more specialized scrapers and choosing which one(s) is/are the more appropriate for the job.