Add a copy of \Illuminate\Console\Command
This class is a copy of the latest 4.2 version of \Illuminate\Console\Command.

This hack is required by the fact that, starting with version 5, the illuminate/console package breaks standalone command line tools that are independent from a Laravel project. But, at the same time, we want methuselah to be able to work within projects using Laravel 5 (and thus requiring version 5.* of illuminate/console), not only with Laravel 4.*. Hence this hacky solution.